fix: #WB-2267, set dayjs locale globally
