+++ packages/react/src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from "./functions";
 export * from "./MockedDataProvider";
 export * from "./ref";
-export * from "./date";
 export * from "./time";
 export * from "./video";
